requirements.txt
numpy==1.16.2
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyserial==3.4
pytz==2018.9
